% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colourscale.R
\name{add_colourscale}
\alias{add_colourscale}
\title{Add scale to frames}
\usage{
add_colourscale(frames, type, colours, labels = waiver(),
  legend_title = NULL, verbose = TRUE)
}
\arguments{
\item{frames}{list of \code{ggplot} objects, crated with \code{\link{create_frames}}.}

\item{type}{character, either \code{"gradient"} or \code{"discrete"}. Must be equal to the defintion of argument \code{r_type} with which \code{frames} have been created (see \code{\link{create_frames}}).}

\item{colours}{character, a vector of colours. If \code{type = "categorcial"}, number of colours must be equal to the number of classes contained in the raster imagery with which \code{frames} have been created.}

\item{labels}{character, a vector of labels with the same length as \code{colours}. Ignored, if \code{type = "gradient"}.}

\item{legend_title}{character, a legend title.}

\item{verbose}{logical, if \code{TRUE}, messages on the function's progress are displayed (default).}
}
\value{
List of frames.
}
\description{
This function adjust the colour scales of the animation frames created with \code{\link{create_frames}}.
}
\seealso{
\link{create_frames}
}
\author{
Jakob Schwalb-Willmann
}
