% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to frames}
\usage{
add_labels(frames, title = waiver(), subtitle = waiver(),
  caption = waiver(), tag = waiver(), x = waiver(), y = waiver(),
  verbose = TRUE)
}
\arguments{
\item{frames}{list of \code{ggplot} objects, crated with \code{\link{create_frames}}.}

\item{title}{character, frame title.}

\item{subtitle}{character, frame subtitle.}

\item{caption}{character, frame caption.}

\item{tag}{character, frame tag.}

\item{x}{character, label of the x axis.}

\item{y}{character, label of the y axis.}
}
\value{
List of frames.
}
\description{
This function adds character labels such as title or axis labels to animation frames created with \code{\link{create_frames}}.
}
\seealso{
\link{create_frames}
}
\author{
Jakob Schwalb-Willmann
}
