% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_northarrow.R
\name{add_northarrow}
\alias{add_northarrow}
\title{Add visual elements to animation frames}
\usage{
add_northarrow(frames, ..., verbose = TRUE)
}
\value{
List of frames.
}
\description{
Functions that add visual elements to animation frames created with \code{links{create_frames}} to adjust the appearance of the movement animation before calling \code{\link{animate_frames}}.
}
\details{
\code{add_labels} adds title, sub title and axis labels to all frames
}
\seealso{
\link{create_frames}
}
\author{
Jakob Schwalb-Willmann
}
