% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_progress.R
\name{add_progress}
\alias{add_progress}
\title{Add progress bar to frames}
\usage{
add_progress(frames, colour = "grey", size = 1.8, y_nudge = 0, verbose = TRUE)
}
\arguments{
\item{frames}{an object of class \code{moveVis} created with \code{\link{frames_spatial}}.}

\item{colour}{character, progress bar colour.}

\item{size}{numeric, progress bar line size (line width).}

\item{y_nudge}{numeric, amount of vertical distance to move the progress bar (default: top of plot area).}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}
}
\value{
A frames object of class \code{moveVis}.
}
\description{
This function adds a progress bar to frames created with \code{\link{frames_spatial}}.
}
\examples{
library(moveVis)
library(move2)
library(terra)

data("move_data", package = "moveVis")
r <- readRDS(example_data(file = "basemap_data.rds"))

# align movement
m <- align_move(move_data, res = units::set_units(4, "min"))

# create frames and add northarrow
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\% 
  add_progress()
frames[[100]]

# or in white at another position
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\% 
  add_progress(colour = "red", size = 2.5)
frames[[100]]

}
\seealso{
\code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{animate_frames}}
}
\author{
Jakob Schwalb-Willmann
}
