% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scalebar.R
\name{add_scalebar}
\alias{add_scalebar}
\title{Add visual elements to animation frames}
\usage{
add_scalebar(frames, verbose = TRUE)
}
\arguments{
\item{frames}{list of \code{ggplot} objects, crated with \code{\link{create_frames}}.}

\item{verbose}{logical, if \code{TRUE}, messages on the function's progress are displayed (default).}

\item{title}{character, frame title.}

\item{subtitle}{character, frame subtitle.}

\item{caption}{character, frame caption.}

\item{tag}{character, frame tag.}

\item{x}{character, label of the x axis.}

\item{y}{character, label of the y axis.}
}
\value{
List of frames.
}
\description{
Functions that add visual elements to animation frames created with \code{links{create_frames}} to adjust the appearance of the movement animation before calling \code{\link{animate_frames}}.
}
\details{
\code{add_labels} adds title, sub title and axis labels to all frames
}
\seealso{
\link{create_frames}
}
\author{
Jakob Schwalb-Willmann
}
