% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{add_text}
\alias{add_text}
\title{Add text to frames}
\usage{
add_text(
  frames,
  labels,
  x,
  y,
  colour = "black",
  size = 3,
  type = "text",
  verbose = TRUE
)
}
\arguments{
\item{frames}{an object of class \code{moveVis} created with \code{\link{frames_spatial}}.}

\item{labels}{character, text to be added to frames. Either a single character value or a character vector of same length as \code{frames}.}

\item{x}{numeric, position of text on the x scale. Either a single numeric value or a numeric vector of same length as \code{frames}.}

\item{y}{numeric, position of text on the y scale. Either a single numeric value or a numeric vector of same length as \code{frames}.}

\item{colour}{character, the text colour(s). Either a single character value or a character vector of same length as \code{frames}.}

\item{size}{numeric, the text size(s). Either a single numeric value or a numeric vector of same length as \code{frames}.}

\item{type}{character, either \code{"text"} to draw text or \code{"label"} to draw text inside a box.}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}
}
\value{
A frames object of class \code{moveVis}.
}
\description{
This function adds text to frames created with \code{\link{frames_spatial}}.
}
\examples{
library(moveVis)
library(move2)
library(terra)

data("move_data", package = "moveVis")
r <- readRDS(example_data(file = "raster_NDVI.rds"))

# align movement
m <- align_move(move_data, res = units::set_units(4, "min"))

# create frames and add text somewhere to all frames:
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\%
  add_text("Water area", x = 8.959, y = 47.7305, colour = "white", size = 3)
frames[[100]]

# or use the ggplot2 "label" type:
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\%
  add_text("Water area", x = 8.959, y = 47.7305, colour = "black", size = 3, 
           type = "label")
frames[[100]]

}
\seealso{
\code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{animate_frames}}
}
\author{
Jakob Schwalb-Willmann
}
