% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_timestamps.R
\name{add_timestamps}
\alias{add_timestamps}
\title{Add timestamps to frames}
\usage{
add_timestamps(
  frames,
  x = NULL,
  y = NULL,
  format = "\%Y-\%m-\%d \%H:\%M:\%S",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{frames}{an object of class \code{moveVis} created with \code{\link{frames_spatial}}.}

\item{x}{numeric, optional, position of timestamps on the x scale. By default, timestamps will be displayed in the top center.}

\item{y}{numeric, optional, position of timestamps on the y scale.}

\item{format}{character, optional, format of timestamps to be displayed in, passed to \code{\link{strftime}}).}

\item{...}{optional, arguments passed to \code{\link{add_text}}, such as \code{colour}, \code{size}, \code{type}.}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}
}
\value{
A frames object of class \code{moveVis}.
}
\description{
This function adds timestamps to frames created with \code{\link{frames_spatial}}.
}
\examples{
library(moveVis)
library(move2)
library(terra)

data("move_data", package = "moveVis")
r <- readRDS(example_data(file = "raster_NDVI.rds"))

# align movement
m <- align_move(move_data, res = units::set_units(4, "min"))

# create frames and add timestamps as text
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\%
  add_timestamps(type = "text")
frames[[100]]

# or use the ggplot2 "label" type:
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\%
  add_timestamps(type = "label")
frames[[100]]

}
\seealso{
\code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{animate_frames}}
}
\author{
Jakob Schwalb-Willmann
}
