% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_move.R
\name{align_move}
\alias{align_move}
\title{Align movement data}
\usage{
align_move(
  m,
  res = "minimum",
  unit = NA,
  spaceMethod = "greatcircle",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{m}{\code{move} or \code{moveStack}, which is allowed to contain irregular timestamps and diverging temporal resolutions to be aligned (see \code{\link{df2move}} to convert a \code{data.frame} to a \code{move} object).}

\item{res}{either numeric, representing the temporal resolution, to which \code{m} should be aligned to (see argument \code{unit}), or character:
\itemize{
  \item \code{"minimum"} to use the smallest temporal resolution of \code{m} (default)
  \item \code{"maximum"} to use the largest temporal resolution of \code{m}
  \item \code{"mean"} to use the rounded average temporal resolution of \code{m}
}}

\item{unit}{character, temporal unit of \code{res}. Either \code{"secs"}, \code{"mins"}, \code{"hours"}, \code{"days"}. No effect, if \code{res} is not defined.}

\item{spaceMethod}{character, either \code{"euclidean"}, \code{"greatcircle"} or \code{"rhumbline"}, indicating the interpolation function to be used to interpolate locations of \code{m} to the aligned time scale. Interpolation is performed using \code{move::interpolateTime}.}

\item{...}{deprecated arguments including \code{digit}.}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}
}
\value{
Aligned \code{move} or \code{moveStack}, ready to be used with \code{\link{frames_spatial}}-
}
\description{
This function aligns movement data to a uniform time scale with a uniform temporal resolution throughout the complete movement sequence. 
This prepares the provided movement data to be interpretable by \code{\link{frames_spatial}}, which necessitates a uniform time scale and
a consistent, unique temporal resolution for all moving individuals to turn recording times into frame times.
}
\examples{
library(moveVis)
library(move)
data("move_data")

# the tracks in move_data have irregular timestamps and sampling rates.
# print unique timestamps and timeLag
unique(timestamps(move_data))
unique(unlist(timeLag(move_data, units = "secs")))

# use align_move to correct move_data to a uniform time scale and lag using interpolation.
# resolution of 4 minutes:
m <- align_move(m = move_data, res = 4, unit = "mins")
unique(unlist(timeLag(m, units = "mins")))

# resolution of 1 hour:
m <- align_move(move_data, res = 1, unit = "hours")
unique(unlist(timeLag(m, units = "hours")))

# resolution of 15 seconds:
m <- align_move(move_data, res = 15, unit = "secs")
unique(unlist(timeLag(m, units = "secs")))

}
\seealso{
\code{\link{df2move}} \code{\link{frames_spatial}} \code{\link{frames_graph}}
}
\author{
Jakob Schwalb-Willmann
}
