% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_frames3D_RGL.R
\name{animate_3D_RGL}
\alias{animate_3D_RGL}
\title{Animate movement on 3D Map}
\usage{
animate_3D_RGL(
  m,
  out_file,
  map_service = "mapbox",
  map_type = "satellite",
  map_token = map_token,
 
    m.crs = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs",
  theta = theta,
  phi = phi,
  sunangle = 45,
  zscale_terrain = 3,
  zscale_movement = 3,
  zoom = 0.8,
  out_ext = "gif",
  ...
)
}
\arguments{
\item{out_file}{character, the output file path, e.g. "/dir/to/".}

\item{map_service}{blindtext}

\item{map_type}{blindtext}

\item{map_token}{blindtext}

\item{m.crs}{blindtext}

\item{theta}{blindtext}

\item{phi}{blindtext}

\item{sunangle}{blindtext}

\item{zscale_terrain}{blindtext}

\item{zscale_movement}{blindtext}

\item{zoom}{blindtext}

\item{out_ext}{blindtext}

\item{...}{additional arguments to be passed to the render function.}
}
\value{
A character vector of supported map types
}
\description{
This function returns every supported map type that can be used as input to the \code{map_type} argument of \code{\link{frames_spatial}}.
}
\examples{


}
\seealso{
\code{\link{frames_spatial}}
}
