% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_frames.R
\name{animate_frames}
\alias{animate_frames}
\title{Animate frames}
\usage{
animate_frames(frames, out_file, framerate = 2, width = 800,
  height = 800, display = TRUE, render_function = NULL, ...)
}
\arguments{
\item{out_file}{character, the output file path, e.g. "/dir/to/file.mov". The file extension must correspond to a known video/image format such as mp4, mkv, mov, flv or gif.}

\item{framerate}{numeric, the number of frames to be displayed per second. Default is 2.}

\item{width}{numeric, width of the output animation in pixels.}

\item{height}{numeric, height of the output animation in pixels.}

\item{display}{logical, whether the animation should be displayed after rendering or not.}

\item{render_function}{function, an optional render function. If undefined (default), an appropriate render function is selected automatically depending on the output file extension and the available system libraries.}

\item{...}{additional arguments to be passed to the render function.}
}
\value{
None or an R graphics window displaying the animation
}
\description{
\code{animate_frames} creates an animation from a list of frames computed with \code{\link{create_frames}}.
}
\seealso{
\code{\link{create_frames}}
}
\author{
Jakob Schwalb-Willmann
}
