% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_frames3D.R
\name{animate_frames3D}
\alias{animate_frames3D}
\title{\code{animate_frames3D} creates an animation from a list of frames computed with \code{\link{frames_spatial}}.}
\usage{
animate_frames3D(
  frames,
  out_file,
  scale = 180,
  multicore = TRUE,
  width = 8,
  height = 6,
  windowsize = c(1500, 1125),
  render.camera,
  overwrite = TRUE,
  out_ext = "gif",
  display = TRUE,
  ...
)
}
\arguments{
\item{frames}{list of \code{ggplot2} objects, crated with \code{\link{frames_spatial}}.}

\item{out_file}{character, the output file path, e.g. "/dir/to/".}

\item{scale}{blindtext}

\item{multicore}{blindtext}

\item{width}{numeric, width of the output animation in pixels.}

\item{height}{numeric, height of the output animation in pixels.}

\item{windowsize}{numeric, windowssize of the image in pixel, e.g. c(1500, 1125)}

\item{render.camera}{dataframe, with theta and phi for values. Theta:  Theta values, in degrees. Phi: Azimuth values, in degrees.}

\item{overwrite}{logical, wether to overwrite an existing file, if \code{out_file} is already present.}

\item{out_ext}{output format, gif or mp4}

\item{display}{logical, whether the animation should be displayed after rendering or not.}

\item{...}{additional arguments to be passed to the render function.}
}
\value{
None or the default image/video viewer displaying the animation
}
\description{
View movements on 3D Map
}
\details{
\code{animate_frames3D} is a simple wrapper that displays movement tracks on an 3D map.

An appropriate render function is selected depending on the file extension in \code{out_file}: For \code{.gif} files, \code{gifski::save_gif} is used, for any other (video) format, \code{av::av_capture_graphics} is used.
}
\examples{

}
\seealso{
\code{\link{frames_spatial}}
}
\author{
Jakob Schwalb-Willmann
}
