% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_data.R
\docType{data}
\name{basemap_data}
\alias{basemap_data}
\title{MODIS NDVI 2013 example data of Lake Constance area, Germany}
\format{List containing two lists of equal lengths: a list of rasters containing 10 MODIS NDVI raster class objects and a list of corresponding POSIXct class timestamp objects.}
\source{
MODIS 2013 freely available data, acsessed using MODIS R package: URL \url{https://cran.r-project.org/package=MODIS}
}
\usage{
basemap_data
}
\description{
Dataset containing five MODIS NDVI scenes of 2013 covering the Lake Constance area.
}
\details{
The example data have been pre-processed and have euqal extents and projections.
}
\keyword{datasets}
