% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.moveVis}
\alias{c.moveVis}
\title{Combining moveVis frames}
\usage{
\method{c}{moveVis}(...)
}
\arguments{
\item{...}{two or more objects of class \code{moveVis}.}
}
\value{
A list of \code{moveVis} frames objects.
}
\description{
Method for combining multiple \code{moveVis} frames objects.
}
