% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{animate_move}
\alias{animate_raster}
\alias{animate_stats}
\alias{get_formats}
\alias{get_libraries}
\title{Deprecated functions}
\usage{
animate_move(...)

animate_raster(...)

animate_stats(...)

get_formats(...)

get_libraries(...)
}
\description{
Several functions are deprecated due to a rewrite of moveVis with version 0.10.
}
\details{
The new version of moveVis makes it much easier to animate movement data and multi-temporal imagery (see \code{?moveVis}). You gain more control about the preprocessing of your movement data as well as the visual customization of each animation frame through a more consequent link of \code{moveVis} to \code{gpplot2}.
}
\note{
To install the old version of moveVis (0.9.9), see \url{https://github.com/16EAGLE/moveVis/releases/tag/v0.9.9}.
}
