% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[.moveVis}
\alias{[.moveVis}
\title{Extract moveVis frames}
\usage{
\method{[}{moveVis}(x, i, ...)
}
\arguments{
\item{x}{an object of class \code{moveVis}.}

\item{i}{numeric, index number or sequence of index numbers of the frame(s) to be extracted.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{moveVis} frames object.
}
\description{
Method for extracting individual frames or a sequence of frames from an \code{moveVis} frames object.
}
