% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames3D_rgl.R
\name{frames_rgl}
\alias{frames_rgl}
\title{Create frames of spatial movement maps for animation}
\usage{
frames_rgl(
  m,
  out_file,
  color = rainbow(15),
  own_terrain = FALSE,
  path_terrain,
  map_service = "mapbox",
  map_type = "satellite",
  map_token = map_token,
 
    m.crs = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs",
  pointsize = 5,
  sunangle = 45,
  zscale_terrain = 3,
  zscale_movement = 3,
  point = TRUE,
  ...
)
}
\arguments{
\item{out_file}{character, the output file path, e.g. "/dir/to/".}

\item{color}{color palette for individuals, default: rainbow (15)}

\item{own_terrain}{TRUE if you want to use your own Terrain Model, default: FALSE}

\item{path_terrain}{character, file path for the own terrain model.}

\item{map_service}{choose a mapservice e.g. Mapbox - more informations basemaps::get_maptypes}

\item{map_type}{choose a maptype e.g. satellite  - more informations basemaps::get_maptypes}

\item{map_token}{add a map token if it is necassary for the mapservice e.g. mapbox}

\item{m.crs}{choose crs, default}

\item{pointsize}{size of each point, default 10}

\item{sunangle}{sunangle for 3D Background Image, default 45}

\item{zscale_terrain}{The ratio between the x and y spacing (which are assumed to be equal) and the z axis. For example, if the elevation levels are in units of 1 meter and the grid values are separated by 10 meters, 'zscale' would be 10. Adjust the zscale down to exaggerate elevation features.Default '10'.}

\item{zscale_movement}{define the height of the points, zscale_terrain=zscale_movement - points have same height as basemap}

\item{point}{TRUE: only points are plotted, FALSE: segments are plotted, Default TRUE}

\item{...}{additional arguments to be passed to the render function.}
}
\value{
A rgl scene object, which can be plotted with animate_frames_rgl
}
\description{
\code{frames_rgl} creates a list of rgl scenes displaying movement. Each object represents a single frame. Each frame can be viewed or modified individually. The returned list of frames can be animated using \code{\link{animate_frames_rgl}}.
}
\examples{

}
\seealso{
\code{\link{frames_spatial}}
}
