% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frametimes.R
\name{get_frametimes}
\alias{get_frametimes}
\title{Get frame times from frames}
\usage{
get_frametimes(frames)
}
\arguments{
\item{frames}{an object of class \code{moveVis} created with \code{\link{frames_spatial}}.}
}
\value{
A POSIXct vector of timestamps representing the time associated with each frame.
}
\description{
This function extracts the timestamps associated with each frame from a \code{moveVis} object created using \code{\link{frames_spatial}} or \code{\link{frames_graph}} and returns them as a vector.
}
\examples{
library(moveVis)
library(move)

data("move_data")
# align movement
m <- align_move(move_data, res = 4, unit = "mins")

\dontrun{
frames <- frames_spatial(m, map_service = "osm", map_type = "watercolor")
frames.ts <- get_frametimes(frames)
print(frames.ts)
}

}
\seealso{
\code{\link{frames_spatial}} \code{\link{frames_graph}}
}
