% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tail.moveVis}
\alias{tail.moveVis}
\alias{head.moveVis}
\title{Return first or last frames of an moveVis frames object}
\usage{
\method{tail}{moveVis}(x, n = 6L, ...)

\method{head}{moveVis}(x, n = 6L, ...)
}
\arguments{
\item{x}{an object of class \code{moveVis}.}

\item{n}{an integer of length up to \code{length(x)}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{moveVis} frames object.
}
\description{
Method for returning \code{n} last or first frames of a \code{moveVis} frames objects.
}
