% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_frames.R
\name{join_frames}
\alias{join_frames}
\title{Join multiple frames lists into a single frames list}
\usage{
join_frames(frames_lists, ..., verbose = T)
}
\arguments{
\item{frames_lists}{list, a list of two or more frames lists that you want to combine. All frames lists contained in \code{frames_lists} must be of equal lengths. The contained \code{ggplot2} objects are passed frame-wise to the \code{plotlist} argument of \code{\link{plot_grid}}.}

\item{...}{Further arguments, specifying the appearance of the joined \code{ggplot2} objects, passed to \code{\link{plot_grid}}. See \code{\link{plot_grid}} for further options.}

\item{verbose}{logical, if \code{TRUE}, messages on the function's progress are displayed (default).}
}
\value{
List of ggplot2 objects, each representing a single frame.
}
\description{
This function side-by-side joins the \code{ggplot2} objects of two or more frames lists of equal lengths into a single plot per frame using \code{\link{plot_grid}}. This is useful if you want to side-by-side combine spatial frames returned by \code{\link{frames_spatial}} with graph frames returned by \code{\link{frames_graph}}.
}
