% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{length.moveVis}
\alias{length.moveVis}
\title{Length of moveVis frames}
\usage{
\method{length}{moveVis}(x)
}
\arguments{
\item{x}{an object of class \code{moveVis}.}
}
\value{
Numeric
}
\description{
Method to get length of \code{moveVis} frames, i.e. number of frames.
}
