% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_data.R
\docType{data}
\name{move_data}
\alias{move_data}
\title{Exemplary simualted movement tracks}
\format{\code{Move} object, as used by the \code{move} package.}
\usage{
move_data
}
\description{
This datasetd contains a \code{Move} object, representing cooridnates and acquisition times of three simulated movement tracks, covering a location nearby Lake of Constance, Germany. Individual names are made up for demonstration purposes.
}
\details{
This object is used by some \code{moveVis} examples and unit tests.
}
\note{
All data contained should only be used for testing \code{moveVis} and are not suitable to be used for analysis or interpretation.
}
\keyword{datasets}
