% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Re-exported magrittr forward pipe operator}
\description{
\code{moveVis} has been designed to work with the \link{magrittr} forward pipe operator \code{\%>\%}, which therefore is re-exported by \code{moveVis}.
}
\details{
For a detailed documentation, follow this link: \code{\link[magrittr]{\%>\%}}
}
