% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.moveVis}
\alias{print.moveVis}
\title{Print moveVis frames}
\usage{
\method{print}{moveVis}(x, ...)
}
\arguments{
\item{x}{an object of class \code{moveVis}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisible, used for its side effect.
}
\description{
Method for printing \code{moveVis} frames. Prints show basic information about the object, including number of frames, extent and more.
}
