% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raster_data}
\alias{raster_data}
\alias{example_data}
\title{Example raster data}
\source{
Simulated based on MODIS (MOD13Q1 NDVI)
}
\usage{
example_data(file = c("raster_NDVI.rds", "raster_classification.rds"))
}
\arguments{
\item{file}{character, file name of the example dataset.}
}
\value{
A \code{SpatRasterDataset} dataset of either continuous or discrete values, including acquisition times as \code{POSIXct} objects.
}
\description{
This function provides two example raster datasets as \code{SpatRasterDataset}, covering the Lake of Constance area, as well as invented dates and times that simulate acquisition times.
}
\details{
The returned \code{SpatRasterDataset} objects are used as example data by some \code{moveVis} code examples and unit tests.

The two available datasets are: 

\itemize{
 \item \code{raster_NDVI.rds} representing simulated NDVI images (continuous values)
 \item \code{raster_classification.rds}, representing an NDVI-derived classification (discrete values). NDVI values from -1 to 0 translate into class 1, from 0 to 0.2 into class 2, from 0.2 to 0.4 into class 3, from 0.4 to 0.8 into class 4 and from 0.8 to 1 into class 5.
}

Acquisition times can be retrieved using \code{terra::time(readRDS(example_data()))}
}
\note{
All data contained should only be used for testing \code{moveVis} and are not suitable to be used for analysis or interpretation.
}
