% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{get_maptypes}
\alias{flush_cache}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basemaps}{\code{\link[basemaps:cache]{flush_cache}}, \code{\link[basemaps:maptypes]{get_maptypes}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

