% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rev.moveVis}
\alias{rev.moveVis}
\title{Reverse moveVis frames}
\usage{
\method{rev}{moveVis}(x)
}
\arguments{
\item{x}{an object of class \code{moveVis}.}
}
\value{
A \code{moveVis} frames object.
}
\description{
Method for reversing the order of frames in an \code{moveVis} frames object.
}
