% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveVis-internal.R
\name{[[.moveVis}
\alias{[[.moveVis}
\title{#' names methods
#' @export
names.moveVis <- function(frames){
  unique(frames$move_data$name)
}
render methods}
\usage{
\method{[[}{moveVis}(frames, ...)
}
\description{
#' names methods
#' @export
names.moveVis <- function(frames){
  unique(frames$move_data$name)
}
render methods
}
