% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_multicore.R
\name{use_multicore}
\alias{use_multicore}
\title{Enable multi-core usage}
\usage{
use_multicore(n_cores = NULL, verbose = TRUE)
}
\arguments{
\item{n_cores}{numeric, optional, number of cores to be used. If not defined, the number of cores will be detected automatically (\code{n-1} cores will be used with \code{n} being the number of cores available).}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}
}
\value{
None. This function is used for its side effects.
}
\description{
This function enables multi-core usage of \code{moveVis} by setting the maximum number of cores to be used.
}
\details{
For most tasks, \code{moveVis} is able to use multiple cores to increase computational times through parallelization. By default, multi-core usage is disabled. This function saves the number of cores that \code{moveVis} should use to the global option \code{"moveVis.n_cores"} that can be printed using \code{getOption("moveVis.n_cores")}.

\code{moveVis} uses the \code{parallel} package for parallelization.
}
\examples{
# enable multi-core usage automatically
use_multicore()

# define number of cores manually
use_multicore(n_cores = 2)

}
