% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whitestork_data}
\alias{whitestork_data}
\alias{df}
\alias{m}
\title{White Stork LifeTrack tracks}
\format{
\itemize{
   \item \code{df} is a \code{data.frame} object
   \item \code{m} is a \code{moveStack} object
}

An object of class \code{MoveStack} with 155173 rows and 3 columns.
}
\usage{
data(whitestork_data)
}
\description{
This dataset contains a \code{data.frame} object, representing coordinates and acquisition times of 15 White Storks, migrating from Lake of Constance, SW Germany, to Africa.
}
\details{
These objects are used by some \code{moveVis} examples and have been included for demonstrational purposes.

The dataset represents a subset of the LifeTrack White Stork dataset by Cheng et al. (2019) and Fiedler et al. (2019), available under the Creative Commons license "CC0 1.0 Universal Public Domain Dedication" on Movebank (doi:10.5441/001/1.ck04mn78/1).
}
\references{
Cheng Y, Fiedler W, Wikelski M, Flack A (2019) "Closer-to-home" strategy benefits juvenile survival in a long-distance migratory bird. Ecology and Evolution. doi:10.1002/ece3.5395 
    
Fiedler W, Flack A, Schäfle W, Keeves B, Quetting M, Eid B, Schmid H, Wikelski M (2019) Data from: Study "LifeTrack White Stork SW Germany" (2013-2019). Movebank Data Repository. doi:10.5441/001/1.ck04mn78
}
\keyword{datasets}
