% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_progress.R
\name{add_progress}
\alias{add_progress}
\title{Add progress bar to frames}
\usage{
add_progress(frames, colour = "grey", size = 1.8, verbose = TRUE)
}
\arguments{
\item{frames}{list of \code{ggplot2} objects, crated with \code{\link{frames_spatial}}.}

\item{verbose}{logical, if \code{TRUE}, messages on the function's progress are displayed (default).}

\item{title}{character, frame title.}

\item{subtitle}{character, frame subtitle.}

\item{caption}{character, frame caption.}

\item{tag}{character, frame tag.}

\item{x}{character, label of the x axis.}

\item{y}{character, label of the y axis.}
}
\value{
List of frames.
}
\description{
This function adds a progress bar to animation frames created with \code{\link{frames_spatial}}.
}
\seealso{
\link{frames_spatial}
}
\author{
Jakob Schwalb-Willmann
}
