% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scalebar.R
\name{add_scalebar}
\alias{add_scalebar}
\title{Add scalebar to frames}
\usage{
add_scalebar(
  frames,
  distance = NULL,
  height = 0.015,
  position = "bottomleft",
  x = NULL,
  y = NULL,
  colour = "black",
  label_margin = 1.2,
  units = "km",
  verbose = TRUE
)
}
\arguments{
\item{frames}{an object of class \code{moveVis} created with \code{\link{frames_spatial}}.}

\item{distance}{numeric, optional. Distance displayed by the scalebar (in either km or miles defined by argument \code{units}) By default, the displayed distance is calculated automatically.}

\item{height}{numeric, height of the scalebar in a range from 0 to 1 as the proportion of the overall height of the frame map. Default is 0.015.}

\item{position}{character, position of the scalebar on the map. Either \code{"bottomleft", "upperleft", "upperright", "bottomright"}. Ignored, if \code{x} and \code{y} are set.}

\item{x}{numeric, position of the bottom left corner of the scalebar on the x axis. If not set, \code{position} is used to calculate the position of the scalebar.}

\item{y}{numeric, position of the bottom left corner of the scalebar on the y axis. If not set, \code{position} is used to calculate the position of the scalebar.}

\item{colour}{character, colour of the distance labels. Default is \code{"black"}.}

\item{label_margin}{numeric, distance of the labels to the scalebar as a proportion of the height of the scalebar (e.g. if set to 2, the labels will be positioned with a distance to the scalebar of twice the scalebar height).}

\item{units}{character, either "km" for kilometers or "miles" for miles.}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}
}
\value{
A frames object of class \code{moveVis}.
}
\description{
This function adds a scalebar to frames created with \code{\link{frames_spatial}}.
}
\details{
For adding more sophisticated map features, I recommend using \code{ggspatial}.
}
\examples{
library(moveVis)
library(move2)
library(terra)

data("move_data", package = "moveVis")
r <- readRDS(example_data(file = "basemap_data.rds"))

# align movement
m <- align_move(move_data, res = units::set_units(4, "min"))

# create frames and add scaleebar
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\%
  add_scalebar()
frames[[100]]

# or in white at another position
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\%
  add_scalebar(colour = "white", position = "bottomright")
frames[[100]]

# or with another height
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\%
  add_scalebar(colour = "white", position = "bottomright", height = 0.025)
frames[[100]]
}
\seealso{
\code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{animate_frames}}
}
\author{
Jakob Schwalb-Willmann
}
