% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{add_text}
\alias{add_text}
\title{Add static or dynamic text to frames}
\usage{
add_text(frames, labels, x, y, colour = "white", size = 3,
  type = "text", verbose = TRUE)
}
\arguments{
\item{frames}{list of \code{ggplot} objects, crated with \code{\link{create_frames}}.}

\item{labels}{character, text to be added to frames. Either a single character value or a character vector of same length as \code{frames}.}

\item{x}{numeric, position of text on the x scale. Either a single numeric value or a numeric vector of same length as \code{frames}.}

\item{y}{numeric, position of text on the y scale. Either a single numeric value or a numeric vector of same length as \code{frames}.}

\item{colour}{character, the text colour(s). Either a single character value or a character vector of same length as \code{frames}.}

\item{size}{numeric, the text size(s). Either a single numeric value or a numeric vector of same length as \code{frames}.}

\item{type}{character, either \code{"text"} to draw text or \code{"label"} to draw text inside a box.}

\item{verbose}{logical, if \code{TRUE}, messages on the function's progress are displayed (default).}
}
\value{
List of frames.
}
\description{
This function adds static or dynamically changing text to the animation frames created with \code{\link{create_frames}}.
}
\seealso{
\link{create_frames}
}
\author{
Jakob Schwalb-Willmann
}
