% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_move.R
\name{align_move}
\alias{align_move}
\title{Align movement data}
\usage{
align_move(m, res = "min", unit.digit = "min", unit = "secs",
  spaceMethod = "greatcircle")
}
\arguments{
\item{m}{\code{move} or \code{moveStack}, which is allowed to contain irregular timestamps and diverging temporal resolutions to be aligned}

\item{res}{either numeric, representing the temporal resolution, to which \code{m} should be aligned to (see argument \code{unit}), or character:
\itemize{
  \item \code{"min"} to use the smalles temporal resolution of \code{m} (default)
  \item \code{"max"} to use the largest temporal resolution of \code{m}
  \item \code{"mean"} to use the rounded average temporal resolution of \code{m}
}}

\item{unit.digit}{either numeric, indicating to which digits of a specifc unit (see argument \code{unit}) the time scale of \code{m} should be aligned (e.g. 0 to align the time scale to second ":00", if \code{unit} is set to \code{secs}), or character:
\itemize{
  \item \code{"min"} to use the smallest digit of the defined \code{unit} (default)
  \item \code{"max"} to use the largest digit of the defined \code{unit}
  \item \code{"mean"} to use the rounded average digit of the defined \code{unit}
}}

\item{unit}{character, either \code{"secs"}, \code{"mins"}, \code{"hours"}, \code{"days"}, indicating the temporal unit, to which \code{res} and \code{unit.digit} are referring.}

\item{spaceMethod}{character, either \code{"euclidean"}, \code{"greatcircle"} or \code{"rhumbline"}, indicating the interpolation function to be used to interpolate locations of \code{m} to the aligned time scale. Interpolation is performed using \code{move::interpolateTime}.}
}
\value{
Aligned \code{move} or \code{moveStack}, ready to be used with \code{\link{create_frames}}-
}
\description{
This function aligns movement data to a uniform time scale with a uniform temporal resolution throughout the complete movement sequence. 
This prepares the provided movement data to be interpretable by \code{\link{create_frames}}, which necessitates a uniform time scale and
a consistent, unique temporal resolution for all moving individuals to turn recording times into frame times.
}
\seealso{
\link{create_frames}
}
\author{
Jakob Schwalb-Willmann
}
