% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_frames.R
\name{animate_frames}
\alias{animate_frames}
\title{Animate frames}
\usage{
animate_frames(
  frames,
  out_file,
  fps = 25,
  width = 700,
  height = 700,
  res = 100,
  end_pause = 0,
  display = TRUE,
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{frames}{an object of class \code{moveVis} created with \code{\link{frames_spatial}}.}

\item{out_file}{character, the output file path, e.g. "/dir/to/file.mov". The file extension must correspond to a file format known by the available renderers of the running system. Use \code{\link{suggest_formats}} to get a vector of suggested known file formats.}

\item{fps}{numeric, the number of frames to be displayed per second. Default is 2.}

\item{width}{numeric, width of the output animation in pixels.}

\item{height}{numeric, height of the output animation in pixels.}

\item{res}{numeric, resolution of the output animation in ppi.}

\item{end_pause}{numeric, defining how many seconds the last frame of the animation should be hold to add a pause at the the end of the animation. Default is 0 seconds to not add a pause.}

\item{display}{logical, whether the animation should be displayed after rendering or not.}

\item{overwrite}{logical, wether to overwrite an existing file, if \code{out_file} is already present.}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}

\item{...}{additional arguments to be passed to the render function.}
}
\value{
None or the default image/video viewer displaying the animation
}
\description{
\code{animate_frames} creates an animation from \code{moveVis} frames computed with \code{\link{frames_spatial}}, \code{\link{frames_graph}} or \code{\link{join_frames}}.
}
\details{
An appropriate render function is selected depending on the file extension in \code{out_file}: For \code{.gif} files, \code{gifski::save_gif} is used, for any other (video) format, \code{av::av_capture_graphics} is used.
}
\examples{
library(moveVis)
library(move2)
library(terra)

data("move_data", package = "moveVis")
r <- readRDS(example_data(file = "basemap_data.rds"))

# align movement
m <- align_move(move_data, res = units::set_units(4, "min"))

# create frames 
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) \%>\% 
  add_colourscale(
    type = "gradient", colours = c("orange", "white", "darkgreen"),
    legend_title = "NDVI") \%>\% 
  add_northarrow(position = "bottomleft") \%>\% 
  add_scalebar(colour = "white", position = "bottomright") \%>\% 
  add_progress() \%>\% 
  add_timestamps(type = "label")
 
# check available formats
suggest_formats()

\dontrun{
# animate frames as GIF
out_file <- tempfile(fileext = ".gif")
animate_frames(frames, out_file = out_file)
browseURL(out_file) # view animation
 
# animate frames as mov
out_file <- tempfile(fileext = ".mov")
animate_frames(frames, out_file = out_file)
browseURL(out_file) # view animation
}
}
\seealso{
\code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{join_frames}}
}
\author{
Jakob Schwalb-Willmann
}
