% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_frames.R
\name{animate_frames}
\alias{animate_frames}
\title{Animate frames}
\usage{
animate_frames(frames, out_file, fps = 25, width = 900, height = 900,
  res = 130, display = TRUE, overwrite = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{frames}{list of \code{ggplot2} objects, crated with \code{\link{frames_spatial}}.}

\item{out_file}{character, the output file path, e.g. "/dir/to/file.mov". The file extension must correspond to a file format known by the available renderers of the running system. Use \code{\link{suggest_formats}} to get a vector of suggested known file formats.}

\item{fps}{numeric, the number of frames to be displayed per second. Default is 2.}

\item{width}{numeric, width of the output animation in pixels.}

\item{height}{numeric, height of the output animation in pixels.}

\item{display}{logical, whether the animation should be displayed after rendering or not.}

\item{verbose}{logical, if \code{TRUE}, messages on the function's progress are displayed (default).}

\item{...}{additional arguments to be passed to the render function.}
}
\value{
None or an R graphics window displaying the animation
}
\description{
\code{animate_frames} creates an animation from a list of frames computed with \code{\link{frames_spatial}}.
}
\details{
An appropriate render function is selected depending on the file extension in \code{out_file}: For \code{.gif} files, \code{gifski::save_gif} is used, for any other (video) format, \code{av::av_capture_graphics} is used.
}
\seealso{
\code{\link{frames_spatial}}
}
\author{
Jakob Schwalb-Willmann
}
