% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_frames3D.R
\name{animate_frames3D}
\alias{animate_frames3D}
\title{\code{animate_frames3D} creates an animation from a list of frames computed with \code{\link{frames_spatial}}.}
\usage{
animate_frames3D(
  frames,
  out_file,
  scale = 180,
  multicore = TRUE,
  width = 8,
  height = 6,
  windowsize = c(1500, 1125),
  render.camera,
  overwrite = TRUE,
  out_ext = "gif",
  ..
)
}
\arguments{
\item{frames}{list of \code{ggplot2} objects, crated with \code{\link{frames_spatial}}.}

\item{out_file}{}

\item{...}{additional arguments to be passed to the render function.}
}
\value{
None or the default image/video viewer displaying the animation
}
\description{
#' View movements on 3D Map
}
\details{

}
\examples{

}
\seealso{
\code{\link{frames_spatial}}
}
\author{
Jakob Schwalb-Willmann
}
