% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basemap_data}
\alias{basemap_data}
\alias{example_data}
\title{Example NDVI data}
\format{
\code{SpatRasterDataset}, consisting of NDVI layers and acquisition times as \code{POSIXct} objects.
}
\source{
Simulated based on MODIS (MOD13Q1 NDVI)
}
\usage{
example_data(file = "basemap_data.rds")
}
\description{
This dataset is a SpatRasterDataset, representing simulated NDVI images covering the Lake of Constance area, 
as well as invented dates and times that simulate acquisition times.
}
\details{
This object is used by some \code{moveVis} examples and unit tests. Use \code{readRDS(example_data())} to retrieve the SpatRasterDataset. 

Times can be retrieved using \code{terra::time(readRDS(example_data()))}
}
\note{
All data contained should only be used for testing \code{moveVis} and are not suitable to be used for analysis or interpretation.
}
