% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_frames.R
\name{create_frames}
\alias{create_frames}
\title{Create movement animation frames from movement data}
\usage{
create_frames(m, token, verbose = TRUE, dir.bm = paste0(tempdir(),
  "/moveVis/basemap"))
}
\arguments{
\item{m}{movement}

\item{token}{mapbox token}
}
\value{
List of ggplot objects, each representing a single frame.
}
\description{
\code{create_frames} creates a list of \code{ggplot} objects of which each represents a single frame. Each frame can be viewed or modified individually. The returned list of frames can be animated using \code{\link{animate_frames}}.
}
\seealso{
\code{\link{animate_frames}}
}
\author{
Jakob Schwalb-Willmann
}
