% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{subset_move}
\alias{df2move}
\alias{animate_move}
\alias{animate_raster}
\alias{animate_stats}
\alias{get_formats}
\alias{get_libraries}
\title{Deprecated functions}
\usage{
subset_move(...)

df2move(...)

animate_move(...)

animate_raster(...)

animate_stats(...)

get_formats(...)

get_libraries(...)
}
\arguments{
\item{...}{deprecated arguments.}
}
\description{
Several functions have been deprecated over the development cycles of moveVis.
}
\note{
To install older versions of moveVis, see \url{https://github.com/16EAGLE/moveVis/releases/} or visit the CRAN archive.
}
\seealso{
\code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{join_frames}} \code{\link{animate_frames}}
}
