% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames_graph.R
\name{frames_graph}
\alias{frames_graph}
\title{Create frames of movement-environemnt interaction graphs for animation}
\usage{
frames_graph(m, r_list, r_times, fade_raster = TRUE,
  return_data = FALSE, graph_type = "flow", path_size = 1.5,
  path_legend = TRUE, path_legend_title = "Names", hist_min = NULL,
  hist_max = NULL, hist_by = 0.1, verbose = T)
}
\arguments{
\item{m}{\code{move} or \code{moveStack} of uniform time scale and time lag, e.g. prepared with \code{\link{align_move}} (recommended). May contain a column named \code{colour} to control path colours (see \code{details}).}

\item{r_list}{list of \code{raster} or \code{rasterStack}. Each list element referrs to the times given in \code{r_times}. Use single-layer \code{raster} objects for gradient or discrete data (see \code{r_type}). Use a  \code{rasterStack} containing three bands for RGB imagery (in the order red, green, blue).}

\item{r_times}{list of \code{POSIXct} times. Each list element represents the time of the corresponding element in \code{r_list}. Must be of same length as \code{r_list}.}

\item{fade_raster}{logical, if \code{TRUE}, \code{r_list} is interpolated over time based on \code{r_times}. If \code{FALSE}, \code{r_list} elements are assigned to those frames closest to the equivalent times in \code{r_times}.}

\item{return_data}{logical, if \code{TRUE}, instead of a list of frames, a \code{data.frame} containing the values extracted from \code{r_list} per individual, location and time is returned. This \code{data.frame} can be used to create your own multi- or monotemporal \code{ggplot2} movement-environemnt interaction graphs.}

\item{graph_type}{character, defines the type of multi-temporal graph that should be drawn as frames. Currently supported graphs are:
\itemize{
   \item \code{"flow"}, a time flow graph with frame time on the x axis and values of the visited cell at x on the y axis per individual track
   \item \code{"hist"}, a cumulative histogram with cell values on the x axis and time-cumulative counts of visits on the y axis per individual track.
}}

\item{path_size}{numeric, size of each path.}

\item{path_legend}{logical, wether to add a path legend from \code{m} or not. Legend tracks and colours will be ordered by the tracks' temporal apperances, not by their order in \code{m}.}

\item{path_legend_title}{character, path legend title. Default is \code{"Names"}.}

\item{verbose}{logical, if \code{TRUE}, messages on the function's progress are displayed (default).}
}
\value{
List of ggplot2 objects, each representing a single frame. If \code{return_data} is \code{TRUE}, a \code{data.frame} is returned (see \code{return_data}).
}
\description{
\code{frames_graph} creates a list of \code{ggplot2} objects of which each represents a single frame. Each frame can be viewed or modified individually. The returned list of frames can be animated using \code{\link{animate_frames}}.
}
\details{
To later on side-by-side join spatial frames created using \code{\link{frames_spatial}} with statistical frames created with \code{\link{create_stats}} for animation,
equal inputs must have been used for both function calls for each of the arguments \code{m}, \code{r_list}, \code{r_times} and \code{fade_raster}.
}
\seealso{
\code{\link{frames_spatial}} \code{\link{animate_frames}}
}
\author{
Jakob Schwalb-Willmann
}
