% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frametimes.R
\name{get_frametimes}
\alias{get_frametimes}
\title{Get frame times from frames}
\usage{
get_frametimes(frames)
}
\arguments{
\item{frames}{an object of class \code{moveVis} created with \code{\link{frames_spatial}}.}
}
\value{
A POSIXct vector of timestamps representing the time associated with each frame.
}
\description{
This function extracts the timestamps associated with each frame from a \code{moveVis} object created using \code{\link{frames_spatial}} or \code{\link{frames_graph}} and returns them as a vector.
}
\examples{
library(moveVis)
library(move2)
library(terra)

data("move_data", package = "moveVis")
r <- readRDS(example_data(file = "basemap_data.rds"))

# align movement
m <- align_move(move_data, res = units::set_units(4, "min"))

# create frames 
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) 

# get frametimes
get_frametimes(frames)

}
\seealso{
\code{\link{frames_spatial}} \code{\link{frames_graph}}
}
