% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maptypes.R
\name{get_maptypes}
\alias{get_maptypes}
\title{Get all supported map types}
\usage{
get_maptypes(map_service = NULL)
}
\arguments{
\item{map_service}{character, optional, either \code{"mapbox"} or \code{"osm"}. Otherwise, a list of map types for both services is returned.}
}
\value{
A character vector of supported map types
}
\description{
This function returns every supported map type that can be used as input to the \code{map_type} argument of \code{\link{frames_spatial}}.
}
\seealso{
\code{\link{frames_spatial}}
}
