% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{move_data}
\alias{move_data}
\title{Example simulated movement tracks}
\format{
\code{move2} object, as used by the \code{move2} package.
}
\usage{
data(move_data)
}
\description{
This dataset contains a \code{move2} object, representing coordinates and acquisition times of three simulated movement tracks, covering a location nearby Lake of Constance, Germany. Individual names are made up for demonstration purposes.
}
\details{
This object is used by some \code{moveVis} examples and unit tests.
}
\note{
All data contained should only be used for testing \code{moveVis} and are not suitable to be used for analysis or interpretation.
}
\keyword{datasets}
