% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{get_maptypes}
\alias{flush_cache}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
  \item{basemaps}{\code{\link[basemaps]{get_maptypes}}, \code{\link[basemaps]{flush_cache}}}
}
}
\keyword{internal}
