% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_frame.R
\name{render_frame}
\alias{render_frame}
\title{Render an individual frame}
\usage{
render_frame(frames, i = length(frames), engine = "ggplot2")
}
\arguments{
\item{frames}{an object of class \code{moveVis} created with \code{\link{frames_spatial}}.}

\item{i}{numeric, index number of the frame to be rendered. Per default, the last frame is rendered.}

\item{engine}{character, the default engine to render and animate frames with. Either \code{"ggplot2"} for 2D or \code{"rgl"} for 3D renders and animations.}
}
\description{
This function renders an individual frame. It yields the same result as if an individual frame is extracted using default subsetting \code{[[]]} with the difference that the rendering engine can be defined as an argument. Currently, \code{moveVis} supports \code{ggplot2} for 2D and \code{rgl} for 3D renders and animations.
}
\examples{

library(moveVis)
library(move)
data("move_data")
data("basemap_data")

r_list <- basemap_data[[1]]
r_times <- basemap_data[[2]]

# align
m <- align_move(m = move_data, res = 4, unit = "mins")

# create frames
frames <- frames_spatial(m, r_list = r_list, r_times = r_times, fade_raster = TRUE)

# viewing frames calling this function:
render_frame(frames) # displays the last frame in 2D using ggplot2 (default)
render_frame(frames, i = 100) # displays frame 100 in 2D using ggplot2 (default)

\dontrun{
render_frame(frames, i = 100, engine = "rgl") # displays frame 100 in 3D using rgl

# alternatively, you can simply use `[[` to do this:
frames[[100]] # displays frame 100 in 2D using ggplot2 (default)

# use the settings to change the default rendering engine
set_engine(engine = "rgl")
frames[[100]] # displays frame 100 in 3D using rgl
}
}
