% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/render_frame.R
\name{[[.moveVis}
\alias{[[.moveVis}
\alias{render_frame}
\title{Render an individual frame}
\usage{
\method{[[}{moveVis}(x, i, ...)

render_frame(frames, i = length(frames))
}
\arguments{
\item{x}{frames as an object of class moveVis.}

\item{i}{numeric, index number of the frame to be rendered.}

\item{...}{additional arguments, currently not used.}

\item{frames}{frames as an object of class moveVis.}
}
\description{
This function renders an individual frame. It yields the same result as if an individual frame is extracted using default subsetting \code{[[]]}.
}
\examples{
library(moveVis)
library(move2)
library(terra)
 
data("move_data", package = "moveVis")
r <- readRDS(example_data(file = "raster_NDVI.rds"))
 
# align movement
m <- align_move(move_data, res = units::set_units(4, "min"))
 
# create spatial frames 
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE)
 
# viewing frames calling this function:
render_frame(frames) # displays the last frame 
render_frame(frames, i = 100) # displays frame 100 
 
# alternatively, you can simply use `[[` to do this:
frames[[100]] # displays frame 100
}
