% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_spatial.R
\name{view_spatial}
\alias{view_spatial}
\title{View movements on an interactive map}
\usage{
view_spatial(
  m,
  render_as = "mapview",
  time_labels = TRUE,
  stroke = TRUE,
  path_colours = NULL,
  colour_paths_by = move2::mt_track_id_column(m),
  path_legend = TRUE,
  path_legend_title = colour_paths_by,
  verbose = TRUE
)
}
\arguments{
\item{m}{\code{move2} object.}

\item{render_as}{character, either \code{'mapview'} to return a \code{mapview} map or \code{'leaflet'} to return a \code{leaflet} map.}

\item{time_labels}{logical, whether to display timestamps for each track fix when hovering it with the mouse cursor.}

\item{stroke}{logical, whether to draw stroke around circles.}

\item{path_colours}{character or palette function to use to colour the tracks in `m`. If a character vector, must be a vector of the same length as the number of levels in the attribute being used to colour the tracks (see argument `colour_paths_by`). If a function, must accept an integer and return a vector of colours of that length. By default, colours are selected from a standard rainbow palette per individual track.}

\item{colour_paths_by}{character indicating the name of an attribute column in `m` to use to define path colours. This attribute can be either an event or track-level attribute. By default, colours by track ID.}

\item{path_legend}{logical, whether to add a path legend from \code{m} or not. When colouring tracks by a qualitative attribute, legend entries will be ordered by the levels of that attribute (if a factor) or alphabetically (if a character).}

\item{path_legend_title}{character, path legend title. Defaults to the column name specified in `colour_paths_by`.}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}
}
\value{
An interactive \code{mapview} or \code{leaflet} map.
}
\description{
\code{view_spatial} is a simple wrapper that displays movement tracks on an interactive \code{mapview} or \code{leaflet} map.
}
\examples{
\dontrun{
library(moveVis)
library(move2)

data("move_data", package = "moveVis")

view_spatial(move_data)
 
# return a leaflet map (leaflet must be installed)
view_spatial(move_data, render_as = "leaflet")
 
# turn off time labels and legend
view_spatial(move_data, time_labels = FALSE, path_legend = FALSE)
}
}
\seealso{
\code{\link{frames_spatial}}
}
\author{
Jakob Schwalb-Willmann
}
